requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
argparse==1.2.1
beautifulsoup4==4.4.0
elasticsearch==1.6.0
elasticsearch-dsl==0.0.5.dev
itsdangerous==0.24
python-dateutil==2.4.2
requests==2.7.0
restless==2.0.1
six==1.9.0
urllib3==1.10.4
wsgiref==0.1.2
